move agent_skills to plugins
